- Added background image for Grid View (needs more work) - Added default background color (light gray) for both List and Grid Views.
