+++ src/qt_gui/game_list_frame.h
@@ -14,6 +14,7 @@
 #include "game_list_grid.h"
 #include "game_list_item.h"
 #include "game_list_table.h"
+#include "game_list_utils.h"
 #include "gui_settings.h"
 
 class GameListFrame : public CustomDockWidget {
@@ -87,6 +88,9 @@ private:
     int m_sort_column;
     QMap<QString, QString> m_titles;
 
+    // Game List Utils
+    GameListUtils m_game_list_utils;
+
     // List Mode
     bool m_is_list_layout = true;
     bool m_old_layout_is_list = true;
@@ -111,39 +115,6 @@ private:
     qreal m_margin_factor;
     qreal m_text_factor;
 
-    QString formatSize(qint64 size) {
-        static const QStringList suffixes = {"B", "KB", "MB", "GB", "TB"};
-        int suffixIndex = 0;
-
-        while (size >= 1024 && suffixIndex < suffixes.size() - 1) {
-            size /= 1024;
-            ++suffixIndex;
-        }
-
-        return QString("%1 %2").arg(size).arg(suffixes[suffixIndex]);
-    }
-
-    QString getFolderSize(const QDir& dir) {
-
-        QDirIterator it(dir.absolutePath(), QDirIterator::Subdirectories);
-        qint64 total = 0;
-
-        while (it.hasNext()) {
-            // check if entry is file
-            if (it.fileInfo().isFile()) {
-                total += it.fileInfo().size();
-            }
-            it.next();
-        }
-
-        // if there is a file left "at the end" get it's size
-        if (it.fileInfo().isFile()) {
-            total += it.fileInfo().size();
-        }
-
-        return formatSize(total);
-    }
-
     void SetTableItem(GameListTable* game_list, int row, int column, QString itemStr) {
         QWidget* widget = new QWidget();
         QVBoxLayout* layout = new QVBoxLayout();
@@ -165,73 +136,5 @@ private:
 
         game_list->setItem(row, column, item);
         game_list->setCellWidget(row, column, widget);
-    }
-
-    QImage BlurImage(const QImage& image, const QRect& rect, int radius) {
-        int tab[] = {14, 10, 8, 6, 5, 5, 4, 3, 3, 3, 3, 2, 2, 2, 2, 2, 2};
-        int alpha = (radius < 1) ? 16 : (radius > 17) ? 1 : tab[radius - 1];
-
-        QImage result = image.convertToFormat(QImage::Format_ARGB32);
-        int r1 = rect.top();
-        int r2 = rect.bottom();
-        int c1 = rect.left();
-        int c2 = rect.right();
-
-        int bpl = result.bytesPerLine();
-        int rgba[4];
-        unsigned char* p;
-
-        int i1 = 0;
-        int i2 = 3;
-
-        const int i2Minusi1 = i2 - i1;
-        const int i1Shifted = i1 << 4;
-        const int alphaOver16 = alpha / 16;
-
-        for (int col = c1; col <= c2; col++) {
-            p = result.scanLine(r1) + col * 4;
-            for (int i = i1; i <= i2; i++)
-                rgba[i] = p[i] << 4;
-
-            p += bpl;
-            for (int j = r1; j < r2; j++, p += bpl)
-                for (int i = i1; i <= i2; i++)
-                    p[i] = (rgba[i] += ((p[i] << 4) - rgba[i]) * alpha / 16) >> 4;
-        }
-
-        for (int row = r1; row <= r2; row++) {
-            p = result.scanLine(row) + c1 * 4;
-            for (int i = i1; i <= i2; i++)
-                rgba[i] = p[i] << 4;
-
-            p += 4;
-            for (int j = c1; j < c2; j++, p += 4)
-                for (int i = i1; i <= i2; i++)
-                    p[i] = (rgba[i] += ((p[i] << 4) - rgba[i]) * alpha / 16) >> 4;
-        }
-
-        for (int col = c1; col <= c2; col++) {
-            p = result.scanLine(r2) + col * 4;
-            for (int i = i1; i <= i2; i++)
-                rgba[i] = p[i] << 4;
-
-            p -= bpl;
-            for (int j = r1; j < r2; j++, p -= bpl)
-                for (int i = i1; i <= i2; i++)
-                    p[i] = (rgba[i] += ((p[i] << 4) - rgba[i]) * alpha / 16) >> 4;
-        }
-
-        for (int row = r1; row <= r2; row++) {
-            p = result.scanLine(row) + c2 * 4;
-            for (int i = i1; i <= i2; i++)
-                rgba[i] = p[i] << 4;
-
-            p -= 4;
-            for (int j = c1; j < c2; j++, p -= 4)
-                for (int i = i1; i <= i2; i++)
-                    p[i] = (rgba[i] += ((p[i] << 4) - rgba[i]) * alpha / 16) >> 4;
-        }
-
-        return result;
     }
 };
